refactor(cli): group daemon into lib/daemon (Move 3, RUSH-2708)
Move daemon.ts (+ test) into new lib/daemon/, depth-fix, repoint 25 consumers
(src+tests+bench + string paths), and fix daemon.test.ts REPO_ROOT depth
(climb 3 not 2 after the move). Behavior-preserving; build clean.
