+++ apps/cli/src/commands/detach.ts
@@ -19,7 +19,7 @@ import { gatherLiveTargets } from './go.js';
 import type { ActiveSession } from '../lib/session/active.js';
 import { killSession } from '../lib/tmux/index.js';
 import { getDefaultSocketPath } from '../lib/tmux/paths.js';
-import { getAgentsInvocation } from '../lib/daemon.js';
+import { getAgentsInvocation } from '../lib/daemon/daemon.js';
 import { captureProcessStartTime } from '../lib/pty-server.js';
 import { writeDetachRecord } from '../lib/session/detached.js';
 import { getLogsDir } from '../lib/state.js';
