+++ apps/cli/src/lib/browser/ipc.ts
@@ -5,10 +5,10 @@ import { IS_WINDOWS, ipcEndpoint } from '../platform/index.js';
 import { getHelpersDir } from '../state.js';
 import { resolveBrowserTaskIdleMs } from '../device-config.js';
 import { BrowserService } from './service.js';
-import { startDaemon, stopDaemon } from '../daemon.js';
+import { startDaemon, stopDaemon } from '../daemon/daemon.js';
 import { getCliVersion } from '../version.js';
 import { compareVersions } from '../agent-spec/primitives.js';
-import { getDaemonLogPath } from '../daemon.js';
+import { getDaemonLogPath } from '../daemon/daemon.js';
 import { resolveCallerIdentity } from './caller-identity.js';
 import { actionable } from './service.js';
 import type { IPCRequest, IPCResponse, RefNodeJson } from './types.js';
